Allow to get high quality pictures.
Prior to this commit the quality of the picture was taken from the video
html component, which in general will lead to pretty low resolution.

Instead of that, we add a video constraint, so if possible the video will
project a stream with height of at least 1080p. We store global height
and width from the selected stream and set the canvas to those when snap
is taken.